+++ cg/meta/upload/nipt/nipt.py
@@ -2,14 +2,18 @@
 import datetime as dt
 import logging
 from pathlib import Path
+import requests
 from typing import Iterable, List, Optional
 
+from requests import Response
+
+import paramiko
 from cg.apps.housekeeper.hk import HousekeeperAPI
 from cg.constants import Pipeline
 from cg.exc import HousekeeperFileMissingError
+from cg.meta.upload.nipt.models import StatinaUploadFiles
 from cg.models.cg_config import CGConfig
 from cg.store import Store, models
-from cg.utils import Process
 from housekeeper.store import models as hk_models
 
 LOG = logging.getLogger(__name__)
@@ -24,10 +28,12 @@ class NiptUploadAPI:
         self.sftp_user: str = config.fluffy.sftp.user
         self.sftp_password: str = config.fluffy.sftp.password
         self.sftp_host: str = config.fluffy.sftp.host
+        self.statina_host: str = config.statina.host
+        self.sftp_port = config.fluffy.sftp.port
+        self.sftp_remote_path = config.fluffy.sftp.remote_path
         self.root_dir = Path(config.housekeeper.root)
         self.housekeeper_api: HousekeeperAPI = config.housekeeper_api
         self.status_db: Store = config.status_db
-        self.process = Process(binary="lftp")
         self.dry_run: bool = False
 
     def set_dry_run(self, dry_run: bool) -> None:
@@ -74,19 +80,21 @@ class NiptUploadAPI:
 
     def upload_to_ftp_server(self, results_file: Path) -> None:
         """Upload the result file to the ftp server"""
-
-        parameters: list = [
-            f"sftp://{self.sftp_user}:{self.sftp_password}@{self.sftp_host}",
-            "-e",
-            f'"cd SciLife_Till_StarLims; put {results_file}; bye"',
-        ]
-
-        self.process.run_command(parameters=parameters, dry_run=self.dry_run)
-
-        if self.process.stderr:
-            LOG.info(f"NIPT Upload stderr:\n{self.process.stderr}")
-        if self.process.stdout:
-            LOG.info(f"NIPT Upload stdout:\n{self.process.stdout}")
+        transport: paramiko.Transport = paramiko.Transport((self.sftp_host, self.sftp_port))
+        LOG.info(f"Connecting to SFTP server {self.sftp_host}")
+        transport.connect(username=self.sftp_user, password=self.sftp_password)
+        sftp: paramiko.SFTPClient = paramiko.SFTPClient.from_transport(transport)
+        LOG.info(
+            f"Uploading file {str(results_file)} to remote path "
+            f"{self.sftp_remote_path}/{results_file.name}"
+        )
+        sftp.put(
+            localpath=str(results_file),
+            remotepath=f"/{self.sftp_remote_path}/{results_file.name}",
+            confirm=False,
+        )
+        sftp.close()
+        transport.close()
 
     def update_analysis_uploaded_at_date(self, case_id: str) -> models.Analysis:
         """Updates analysis_uploaded_at for the uploaded analysis"""
@@ -105,3 +113,36 @@ class NiptUploadAPI:
         analysis_obj.upload_started_at = dt.datetime.now()
 
         return analysis_obj
+
+    def get_statina_files(self, case_id: str) -> StatinaUploadFiles:
+        """Get statina files from from housekeeper."""
+
+        hk_results_file: str = self.get_housekeeper_results_file(
+            case_id=case_id, tags=["nipt", "metrics"]
+        )
+        results_file: Path = self.get_results_file_path(hk_results_file)
+
+        hk_multiqc_file: str = self.get_housekeeper_results_file(
+            case_id=case_id, tags=["nipt", "multiqc-html"]
+        )
+        multiqc_file: Path = self.get_results_file_path(hk_multiqc_file)
+
+        hk_segmental_calls_file: str = self.get_housekeeper_results_file(
+            case_id=case_id, tags=["nipt", "wisecondor"]
+        )
+        segmental_calls_file: Path = self.get_results_file_path(hk_segmental_calls_file)
+
+        return StatinaUploadFiles(
+            result_file=results_file.absolute().as_posix(),
+            multiqc_report=multiqc_file.absolute().as_posix(),
+            segmental_calls=segmental_calls_file.as_posix(),
+        )
+
+    def upload_to_statina_database(self, statina_files: StatinaUploadFiles):
+        """Upload nipt data via rest-API."""
+
+        response: Response = requests.post(
+            url=f"{self.statina_host}/insert/batch", data=statina_files.json(exclude_none=True)
+        )
+
+        LOG.info("nipt output: %s", response.text)